add 'recent published drawings' area
fixes #3
On the homepage this shows a "Recent Published Drawings" area that are clickable just like the examples. This list live-updates whenever someone publishes a new drawing and the server keeps a recent history in an Agent (Showoff.RecentDrawings) so that fresh page loads can initialize the list via an ajax call.
Now that we have live-updating and publishing on the homepage there is no need for the 'watch' page so that has been deleted